producers/requirements.txt
kafka-python==2.0.2
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
schedule==1.2.0
python-dotenv==1.0.0
loguru==0.7.0
retrying==1.3.4
pydantic==2.4.2
aiohttp==3.8.5
aiofiles==23.2.1